refactor(api): satisfy detekt for references (ReturnCount, LargeClass)
Rewrite isHttpUrl with runCatching to keep it within detekt's 2-return
limit, and move the three reference ITs into their own
EventReferencesControllerTest so EventControllerTest stays under the
LargeClass threshold. Behaviour unchanged.
